Excludes unix code from windows build and replaces with stubs
