+++ airflow/proxy/daemon.go
@@ -1,3 +1,5 @@
+//go:build !windows
+
 package proxy
 
 import (
