+++ airflow/proxy/routes.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"syscall"
 	"time"
 
 	"github.com/astronomer/astro-cli/config"
@@ -44,41 +43,6 @@ func routesFilePath() string {
 // lockFilePath returns the path used for the flock-based file lock.
 func lockFilePath() string {
 	return filepath.Join(proxyDirPath(), lockFileName)
-}
-
-// acquireLock acquires an exclusive file lock (flock) with timeout.
-// Returns the lock file which must be passed to releaseLock.
-func acquireLock() (*os.File, error) {
-	if err := os.MkdirAll(proxyDirPath(), dirPermRWX); err != nil {
-		return nil, fmt.Errorf("error creating proxy directory: %w", err)
-	}
-
-	f, err := os.OpenFile(lockFilePath(), os.O_CREATE|os.O_RDWR, filePermRW)
-	if err != nil {
-		return nil, fmt.Errorf("error opening lock file: %w", err)
-	}
-
-	deadline := time.Now().Add(lockTimeout)
-	for {
-		err := syscall.Flock(int(f.Fd()), syscall.LOCK_EX|syscall.LOCK_NB)
-		if err == nil {
-			return f, nil
-		}
-		if time.Now().After(deadline) {
-			f.Close()
-			return nil, fmt.Errorf("timed out waiting for routes lock")
-		}
-		time.Sleep(50 * time.Millisecond) //nolint:mnd
-	}
-}
-
-// releaseLock releases the flock and closes the lock file.
-func releaseLock(f *os.File) {
-	if f == nil {
-		return
-	}
-	syscall.Flock(int(f.Fd()), syscall.LOCK_UN) //nolint:errcheck
-	f.Close()
 }
 
 // readRoutes reads routes from the routes file. Returns empty slice if file doesn't exist.
@@ -122,18 +86,6 @@ func writeRoutes(routes []Route) error {
 		return fmt.Errorf("error renaming routes temp file: %w", err)
 	}
 	return nil
-}
-
-// isPIDAlive checks if a process with the given PID is still running.
-var isPIDAlive = func(pid int) bool {
-	if pid <= 0 {
-		return false
-	}
-	proc, err := os.FindProcess(pid)
-	if err != nil {
-		return false
-	}
-	return proc.Signal(syscall.Signal(0)) == nil
 }
 
 // pruneStaleRoutes removes routes whose owner process is no longer alive.
